# specs/001-fastapi-todo-backend/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: FastAPI Secure Multi-User Todo API
  description: Secure multi-user todo application API with JWT authentication and user isolation
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server
paths:
  /api/tasks:
    get:
      summary: List tasks for authenticated user
      description: Retrieve all tasks owned by the authenticated user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of tasks owned by the user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - Missing or invalid JWT token
        '500':
          description: Internal server error
    post:
      summary: Create a new task
      description: Create a new task owned by the authenticated user
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '200':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request - Invalid input data
        '401':
          description: Unauthorized - Missing or invalid JWT token
        '500':
          description: Internal server error
  /api/tasks/{task_id}:
    get:
      summary: Get a specific task
      description: Retrieve a specific task owned by the authenticated user
      security:
        - BearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the task to retrieve
      responses:
        '200':
          description: Task retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - Missing or invalid JWT token
        '403':
          description: Forbidden - Task does not belong to the user
        '404':
          description: Task not found
        '500':
          description: Internal server error
    put:
      summary: Update a specific task
      description: Update all details of a specific task owned by the authenticated user
      security:
        - BearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the task to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request - Invalid input data
        '401':
          description: Unauthorized - Missing or invalid JWT token
        '403':
          description: Forbidden - Task does not belong to the user
        '404':
          description: Task not found
        '500':
          description: Internal server error
    delete:
      summary: Delete a specific task
      description: Delete a specific task owned by the authenticated user
      security:
        - BearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the task to delete
      responses:
        '204':
          description: Task deleted successfully
        '401':
          description: Unauthorized - Missing or invalid JWT token
        '403':
          description: Forbidden - Task does not belong to the user
        '404':
          description: Task not found
        '500':
          description: Internal server error
  /api/tasks/{task_id}/complete:
    patch:
      summary: Toggle task completion status
      description: Toggle the completion status of a specific task owned by the authenticated user
      security:
        - BearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the task to toggle completion status
      responses:
        '200':
          description: Task completion status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - Missing or invalid JWT token
        '403':
          description: Forbidden - Task does not belong to the user
        '404':
          description: Task not found
        '500':
          description: Internal server error
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the task
          example: 1
        title:
          type: string
          description: Title of the task
          example: "Buy groceries"
        description:
          type: string
          description: Optional description of the task
          example: "Need to buy milk, bread, and eggs"
        completed:
          type: boolean
          description: Whether the task is completed
          example: false
        created_at:
          type: string
          format: date-time
          description: Timestamp when the task was created
          example: "2026-02-01T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Timestamp when the task was last updated
          example: "2026-02-01T10:00:00Z"
      required:
        - id
        - title
        - completed
        - created_at
        - updated_at
    TaskCreate:
      type: object
      properties:
        title:
          type: string
          description: Title of the task (required)
          example: "Buy groceries"
        description:
          type: string
          description: Optional description of the task
          example: "Need to buy milk, bread, and eggs"
      required:
        - title
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          description: Updated title of the task
          example: "Buy groceries"
        description:
          type: string
          description: Updated description of the task
          example: "Need to buy milk, bread, eggs, and fruit"